+++ machinevision/accounts/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
+from django.contrib.auth.models import User
 
-# Register your models here.
+
+class CustomUserAdmin(admin.ModelAdmin):
+  list_display = ('username', 'is_active', 'is_staff')
+  list_display_links = ('username',)
+
+admin.site.unregister(User)
+admin.site.register(User, CustomUserAdmin)