+++ machinevision/accounts/api.py
@@ -25,8 +25,7 @@ class RegisterAPI(generics.GenericAPIView):
             "user": UserSerializer(
                 user,
                 context=self.get_serializer_context()
-            ).data,
-            "token": AuthToken.objects.create(user)[1]
+            ).data
         })