+++ machinevision/frontend/src/components/streaming/WebcamStream/WebcamStream.js
@@ -1,20 +1,26 @@
 import React from 'react';
-// https://www.npmjs.com/package/react-webcam
 import Webcam from 'react-webcam';
 
-// Placeholder component until we implement the stream from
-// Jetson Nano
+import Paper from '@material-ui/core/Paper';
+import Grow from '@material-ui/core/Grow';
+
+import { cameraGrowTimeout } from '../../../constants';
+
 const videoConstraints = {
   width: 640,
   height: 480,
   facingMode: 'user',
 };
 
-const WebcamStream = () => {
+const WebcamStream = ({ camEnabled, grow }) => {
   return (
-    <>
-      <Webcam audio={false} videoConstraints={videoConstraints} />
-    </>
+    <Grow in={grow} timeout={cameraGrowTimeout}>
+      <Paper elevation={4}>
+        {camEnabled && (
+          <Webcam audio={false} videoConstraints={videoConstraints} />
+        )}
+      </Paper>
+    </Grow>
   );
 };